+++ src/debugger/disasm.hpp
@@ -11,17 +11,17 @@ class Disassembler {
         Disassembler(MMU *mmu);
         ~Disassembler();
 
-        void setAddress(uint16_t address);
+        void setAddress(uint32_t address);
         std::vector<std::string> disassemble(int count);
         void disassemble_one();
         void setLinePrepend(int spaces) { line_prepend = spaces; }
 
     private:
-        uint16_t address;
+        uint32_t address;
         uint16_t length;
         MMU *mmu;
         int line_prepend;
         std::vector<std::string> output_buffer;
 
-        uint8_t read_mem(uint16_t address);
+        uint8_t read_mem(uint32_t address);
 };
